requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mysql-connector-python==8.0.32
protobuf==3.20.3
pydantic==1.10.6
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
